Fixed bug where player or zombie would get stuck. It was caused by the FreezeTowers. Also brightened the colors for random color generation for players.

+++ src/triGame/game/entities/zombies/Zombie.java
@@ -32,7 +32,7 @@ public class Zombie extends Entity {
 	private int lastFindPathTick = 0;
 	private int tickCount = 0;
 	
-	private int realSpeed = speed;
+	private double realSpeed = speed;
 	private double damage = -100; //damage per second
 	private Path path;
 	private Point lastTargetBlock = new Point(0, 0);
